chore(sunco): refactor file structure and remaining typings
