+++ integrations/sunco/src/types.ts
@@ -23,3 +23,7 @@ export type AckFunction = bp.AnyAckFunction
 export type CreateMessageInput = Parameters<Client['createMessage']>[0]
 export type CreateMessageInputType = CreateMessageInput['type']
 export type CreateMessageInputPayload = CreateMessageInput['payload']
+
+// Channel message payload types
+export type Choice = bp.channels.channel.choice.Choice
+export type Carousel = bp.channels.channel.carousel.Carousel
